Separated intersection.hh to 2 files
